Changes as requested by A Kovacs on 4/8/11
Please make the "name" field into "first name" and "last name" and make them both required.

Please adjust the sales info on the "Sale" page to include the multiple quantities when ordering with one sku.  This can wait for the shopping cart functionality if it's easier.

Move Special Instructions box up higher on sales page, so more folks utilize that instead of the PayPal instructions.  Also, we should consider removing the option in PayPal.